+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## [Unreleased]
+### Added
+- Video downloading
 
 ## [0.1.0] - 2021-07-30
 ### Added